Add conversion to SCF for
Simple matcher of while corresponding to scf.for (the same simple
structure was used in both tests so easy starting point) and with
generating an undefined dialect op (this needs a change upstream in
core). Need to clean up and add tests for it and then upstream to
mlir-hlo repo and then remove here.

emitc-opt --convert-mhlo-while-to-scf /testCorrectGroundTruthWithHMC_canon_inline.mlir -allow-unregistered-dialect

Not the prettiest and requires building HLO dialect until parts are
upstreamed.