Implement MHLO constant to std dialect conversion
ConstantOp of the standard dialect can be use to emit C++ code via the
Cpp printer. We only want to convert the MHLO constant op to the
standard dialect, as other ops are directly converted from MHLO to
EmitC. Anyway, we should consider to add a ConstOp to EmitC itself.