Create initial C target
The C target is a special form of C++ target, which supports emitting C
code for a limited subset of operations and types. For example the C
target does not support tensors or multiple return values. Furthermore,
assignments use = and never the curly-brace-delimited initializer which
is used in the C++ target.