Remove upstreamed EmitC dialect (#166)
Updates LLVM usage to match llvm/llvm-project@876de062. This commit adds
EmitC to the LLVM monorepo, which was accepted to MLIR core [1]. Hence,
the EmitC dialect is removed from the mlir-emitc repository. For now the
Cpp target as well as the conversions remain in our out-of-tree repo.
Further updates the MLIR-HLO submodule to tensorflow/mlir-hlo@05ad9ddf.

[1] https://reviews.llvm.org/D103969.