Add Copyright Notices (#355)
Add copyright notices to the reference implementation for the case of
re-licensing.

As part of https://reviews.llvm.org/D146483, part of the reference
implementation is forked. Even if not as part of this pull request, it
might be of interest to change the license from `Apache-2.0` to
`Apache-2.0 WITH LLVM-exception`. The copyright notices can help to
facilitate this.