Add .get() method to Tensor<...> (#361)
This is used in the MGLO LLVM integration of emitc'ed models. Our
infrastructure internally uses T* buffers to communicate between LLVM
and the model (this was used to abstract away the details of TF/TFLite
runtime vs interacting with a model in a python host vs EmitC) and so we
need to be able to extract the underlying buffer of a EmitC Tensor.

This places a (light) data layout restriction on the Tensor
implementation, as we need it to be in a columnar layout. This is
already what the implementation did, so there is nothing too surprising
here. I have added a unit test to enforce this restriction on the
underlying buffer layout.